Move files around to new packages
