+++ commons/src/main/kotlin/no/ntnu/beardblaster/commons/lobby/AbstractLobbyRepository.kt
@@ -1,7 +1,8 @@
-package no.ntnu.beardblaster.commons
+package no.ntnu.beardblaster.commons.lobby
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
+import no.ntnu.beardblaster.commons.State
 import no.ntnu.beardblaster.commons.game.Game
 import no.ntnu.beardblaster.commons.game.GamePlayer
